Send notification when selected tab is clicked (macOS 10.15+)
When multiple tabbars exist (e.g. a split view, each with its own tabbar), the user may want to switch between the tab bars. The default behavior is that when the user selects the opposite tabbar by clicking its selected tab, that tabbar does not trigger any events, unless a 'click through' (long click) is performed. At times we may want to be notified that a tabbar was selected when the user 'selects' the tab bar by clicking on its tab.

This change sends a DidClickAlreadySelectedTab notification when the user clicks on the 'already selected tab'. You can intercept this notification from the Notification Center and act on it as desired, for example to chnage the first responder. Userinfo stores  references to the tabBarView and the tabBarViewItem.